bullseye/bullseye-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
fastapi[all]
requests